Move ConfigurationPostParams to Config ObjectModel
