+++ src/Service/Controllers/ConfigurationController.cs
@@ -135,32 +135,4 @@ namespace Azure.DataApiBuilder.Service.Controllers
             return BadRequest();
         }
     }
-    /// <summary>
-    /// The required parameters required to configure the runtime.
-    /// </summary>
-    /// <param name="Configuration">The runtime configuration.</param>
-    /// <param name="Schema">The GraphQL schema. Can be left empty for SQL databases.</param>
-    /// <param name="ConnectionString">The database connection string.</param>
-    /// <param name="AccessToken">The managed identity access token (if any) used to connect to the database.</param>
-    /// <param name="Database"> The name of the database to be used for Cosmos</param>
-    public record class ConfigurationPostParameters(
-        string Configuration,
-        string? Schema,
-        string ConnectionString,
-        string? AccessToken)
-    { }
-
-    /// <summary>
-    /// The required parameters required to configure the runtime.
-    /// </summary>
-    /// <param name="Configuration">The runtime configuration.</param>
-    /// <param name="ConfigurationOverrides">Configuration parameters that override the options from the Configuration file.</param>
-    /// <param name="Schema">The GraphQL schema. Can be left empty for SQL databases.</param>
-    /// <param name="AccessToken">The managed identity access token (if any) used to connect to the database.</param>
-    public record class ConfigurationPostParametersV2(
-        string Configuration,
-        string ConfigurationOverrides,
-        string? Schema,
-        string? AccessToken)
-    { }
 }
